refactor(email): Move email logic to lib
