+++ src/lib/email/constants.ts
