+++ src/lib/email/mailHandler.ts
@@ -1,5 +1,5 @@
 import '@pn-server-only'
-import { TRANSPORT_OPTIONS } from './constants'
+import { TRANSPORT_OPTIONS } from '@/lib/email/constants'
 import nodemailer from 'nodemailer'
 import type SMTPPool from 'nodemailer/lib/smtp-pool'
 import type SMTPTransport from 'nodemailer/lib/smtp-transport'
