+++ src/lib/email/systemMail/verifyEmail.tsx
@@ -1,7 +1,7 @@
 import '@pn-server-only'
-import { emailValidationExpiration } from './ConfigVars'
-import { VerifyEmailTemplate } from '@/services/notifications/email/templates/verifyEmail'
-import { sendSystemMail } from '@/services/notifications/email/send'
+import { emailValidationExpiration } from './constants'
+import { VerifyEmailTemplate } from '@/lib/email/templates/verifyEmail'
+import { sendSystemMail } from '@/lib/email/send'
 import { generateJWT } from '@/jwt/jwt'
 import { userSchemas } from '@/services/users/schemas'
 import type { UserFiltered } from '@/services/users/types'
