+++ src/services/auth/operations.ts
@@ -2,7 +2,7 @@ import { authAuth } from './auth'
 import { authSchemas } from './schemas'
 import { userFilterSelection } from '@/services/users/constants'
 import { userSchemas } from '@/services/users/schemas'
-import { sendResetPasswordMail } from '@/services/notifications/email/systemMail/resetPassword'
+import { sendResetPasswordMail } from '@/lib/email/systemMail/resetPassword'
 import { defineOperation } from '@/services/serviceOperation'
 import { ServerError } from '@/services/error'
 import { userOperations } from '@/services/users/operations'
