+++ src/services/cabin/booking/operations.ts
@@ -8,7 +8,7 @@ import { cabinProductPriceIncluder } from '@/services/cabin/product/constants'
 import { defineOperation, defineSubOperation } from '@/services/serviceOperation'
 import { ServerError } from '@/services/error'
 import { cabinReleasePeriodOperations } from '@/services/cabin/releasePeriod/operations'
-import { sendSystemMail } from '@/services/notifications/email/send'
+import { sendSystemMail } from '@/lib/email/send'
 import { notificationOperations } from '@/services/notifications/operations'
 import { cmsParagraphOperations } from '@/cms/paragraphs/operations'
 import { BookingType } from '@/prisma-generated-pn-types'
