+++ src/services/events/registration/operations.ts
@@ -5,7 +5,7 @@ import { eventRegistrationSchemas } from './schemas'
 import { Smorekopp } from '@/services/error'
 import { imageOperations } from '@/services/images/operations'
 import { notificationOperations } from '@/services/notifications/operations'
-import { sendSystemMail } from '@/services/notifications/email/send'
+import { sendSystemMail } from '@/lib/email/send'
 import { userFilterSelection } from '@/services/users/constants'
 import { defineOperation } from '@/services/serviceOperation'
 import { z } from 'zod'
