+++ src/services/notifications/operations.ts
@@ -4,8 +4,8 @@ import { notificationAuth } from './auth'
 import { notificationSchemas } from './schemas'
 import { allNotificationMethodsOn, notificationMethodsArray } from './constants'
 import { availableNotificationMethodIncluder } from './channel/constants'
-import { sendMail } from './email/send'
 import { emailSchemas } from './email/schemas'
+import { sendMail } from '@/lib/email/send'
 import { userFilterSelection } from '@/services/users/constants'
 import { defineOperation, defineSubOperation } from '@/services/serviceOperation'
 import { SpecialNotificationChannel } from '@/prisma-generated-pn-types'
