+++ src/services/users/operations.ts
@@ -10,9 +10,9 @@ import { imageOperations } from '@/services/images/operations'
 import { notificationSubscriptionOperations } from '@/services/notifications/subscription/operations'
 import { readMembershipsOfUser } from '@/services/groups/memberships/read'
 import { NTNUEmailDomain } from '@/services/mail/mailAddressExternal/ConfigVars'
-import { sendVerifyEmail } from '@/services/notifications/email/systemMail/verifyEmail'
+import { sendVerifyEmail } from '@/lib/email/systemMail/verifyEmail'
 import { updateUserOmegaMembershipGroup } from '@/services/groups/omegaMembershipGroups/update'
-import { sendUserInvitationEmail } from '@/services/notifications/email/systemMail/userInvitivation'
+import { sendUserInvitationEmail } from '@/lib/email/systemMail/userInvitivation'
 import { readOmegaMembershipGroup } from '@/services/groups/omegaMembershipGroups/read'
 import { defineOperation } from '@/services/serviceOperation'
 import { readPageInputSchemaObject } from '@/lib/paging/schema'
